Stack watcher changes after metadata scheduler
